Move table header to own component
