+++ src/UI/Table/Table.svelte
@@ -1,17 +1,9 @@
 <script>
   import { createEventDispatcher } from 'svelte';
   import { formatDuration } from '../../Helpers/formatting';
-  import {
-    getWeekNumbers,
-    getStatuses,
-    getYears,
-  } from '../../Helpers/logDataLogic';
   import { getDay, getYear } from '../../Helpers/utils';
-  import { filterData } from '../../Helpers/tableLogic/filter';
   import TableButton from '../Buttons/TableButton.svelte';
-  import TableDropDown from '../Inputs/TableDropdown.svelte';
-  import { routes } from '../../constants';
-  import TableDropdown from '../Inputs/TableDropdown.svelte';
+  import TableHeader from './TableHeader.svelte';
 
   export let logData;
 
@@ -19,104 +11,15 @@
 
   let filteredLogData = logData;
 
-  let filters = {
-    routesFrom: ['no sorting', ...routes],
-    routeFrom: 'no sorting',
-    routesTo: ['no sorting', ...routes],
-    routeTo: 'no sorting',
-    sortDatesOptions: ['no sorting', 'ascending', 'descending'],
-    sortDatesOption: 'no sorting',
-    sortTravelTimeOptions: ['no sorting', 'ascending', 'descending'],
-    sortTravelTimeOption: 'no sorting',
-    statuses: ['all', ...getStatuses(filteredLogData)],
-    status: 'all',
-    weeks: ['all', ...getWeekNumbers(filteredLogData)],
-    week: 'all',
-    years: ['all', ...getYears(filteredLogData)],
-    year: 'all',
-  };
-
-  function resetFilters() {
-    filters.routeFrom = 'no sorting';
-    filters.routeTo = 'no sorting';
-    filters.sortDatesOption = 'no sorting';
-    filters.sortTravelTimeOption = 'no sorting';
-    filters.status = 'all';
-    filters.week = 'all';
-    filters.year = 'all';
-    filteredLogData = logData;
-  }
-
-  function updateData(event, dropdown) {
-    filters[dropdown] = event.target.value;
-    filteredLogData = filterData(filteredLogData, filters[dropdown], dropdown);
+  function doUpdate(data) {
+    filteredLogData = data;
   }
 </script>
 
 <div class="dashboard__container margin-bottom">
-  <div id="reset-btn">
-    <TableButton on:click={resetFilters}>Reset Filters</TableButton>
-  </div>
   <div class="table__container">
     <table>
-      <tr>
-        <th>
-          <TableDropdown
-            label="year"
-            options={filters.years}
-            value={filters.year}
-            on:change={(event) => updateData(event, 'year')}
-          />
-        </th>
-        <th>
-          <TableDropdown
-            label="Week#"
-            options={filters.weeks}
-            value={filters.week}
-            on:change={(event) => updateData(event, 'week')}
-          />
-        </th>
-        <th>
-          <TableDropdown
-            label="Date"
-            options={filters.sortDatesOptions}
-            value={filters.sortDatesOption}
-            on:change={(event) => updateData(event, 'sortDate')}
-          />
-        </th>
-        <th>
-          <TableDropdown
-            label="Working"
-            options={filters.statuses}
-            value={filters.status}
-            on:change={(event) => updateData(event, 'status')}
-          />
-        </th>
-        <th>
-          <TableDropdown
-            label="From"
-            options={filters.routesFrom}
-            value={filters.routeFrom}
-            on:change={(event) => updateData(event, 'routeFrom')}
-          />
-        </th>
-        <th>
-          <TableDropdown
-            label="To"
-            options={filters.routesTo}
-            value={filters.routeTo}
-            on:change={(event) => updateData(event, 'routeTo')}
-          />
-        </th>
-        <th>
-          <TableDropdown
-            label="Travel Time"
-            options={filters.sortTravelTimeOptions}
-            value={filters.sortTravelTimeOption}
-            on:change={(event) => updateData(event, 'sortTravelTime')}
-          />
-        </th>
-      </tr>
+      <TableHeader {doUpdate} {filteredLogData} {logData} />
       {#each filteredLogData as logDate}
         <tr>
           <td>{getYear(logDate.date)}</td>
@@ -150,10 +53,6 @@
 </div>
 
 <style>
-  #reset-btn {
-    margin-bottom: 0.5rem;
-  }
-
   .table__container {
     width: 1150px;
     font-family: 'Helvetica Neue';
@@ -165,11 +64,6 @@
   table {
     margin: auto;
     border-spacing: collapse;
-  }
-
-  th {
-    color: rgba(170, 170, 170, 0.3);
-    text-align: center;
   }
 
   td {
